[ Refactor ] 파일 정리
